feat: edit structure
